nets-builder/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
aiosqlite>=0.19
python-dotenv>=1.0
httpx>=0.27
google-genai>=0.3
